+++ tests/extensions/test_label.py
@@ -20,6 +20,8 @@ from pystac.extensions.label import (
 )
 from pystac.utils import get_opt
 from tests.utils import TestCases, assert_to_from_dict
+from tests.conftest import get_data_file
+import pytest
 
 
 class LabelTypeTest(unittest.TestCase):
@@ -576,3 +578,27 @@ class LabelSummariesTest(unittest.TestCase):
             LabelExtension.ext,
             object(),
         )
+
+
+@pytest.fixture
+def ext_item() -> pystac.Item:
+    ext_item_uri = get_data_file("label/label-example-1.json")
+    return pystac.Item.from_file(ext_item_uri)
+
+
+def test_older_extension_version(ext_item: pystac.Item) -> None:
+    old = "https://stac-extensions.github.io/label/v1.0.0/schema.json"
+    new = "https://stac-extensions.github.io/label/v1.0.1/schema.json"
+
+    stac_extensions = set(ext_item.stac_extensions)
+    stac_extensions.remove(new)
+    stac_extensions.add(old)
+    item_as_dict = ext_item.to_dict(include_self_link=False, transform_hrefs=False)
+    item_as_dict["stac_extensions"] = list(stac_extensions)
+    item = pystac.Item.from_dict(item_as_dict)
+    assert LabelExtension.has_extension(item)
+    assert old in item.stac_extensions
+
+    migrated_item = pystac.Item.from_dict(item_as_dict, migrate=True)
+    assert LabelExtension.has_extension(migrated_item)
+    assert new in migrated_item.stac_extensions