+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Include a copy of the `fields.json` file (for summaries) with each distribution of PySTAC ([#1045](https://github.com/stac-utils/pystac/pull/1045))
 - Make Catalog, Collection `.get_assets()` return a deepcopy ([#1087](https://github.com/stac-utils/pystac/pull/1087))
 - Removed documentation references to `to_dict` methods returning JSON ([#1074](https://github.com/stac-utils/pystac/pull/1074))
+- Expand support for previous extension schema URIs ([#1091](https://github.com/stac-utils/pystac/pull/1091))
 
 ### Deprecated